change relationships of entities: tags <- n:m -> algos/impls rename tag fields
