+++ org.planqk.atlas.core/org.planqk.atlas.core.model/src/main/java/org/planqk/atlas/core/model/AlgorOrImpl.java
@@ -52,10 +52,6 @@ public abstract class AlgorOrImpl extends HasId {
     @OneToMany(cascade = CascadeType.PERSIST)
     private List<Parameter> outputParameters;
 
-    @OneToMany
-    @Setter
-    private List<Tag> tags;
-
     @Getter
     @Setter
     @Type(type = "jsonb")
@@ -79,13 +75,5 @@ public abstract class AlgorOrImpl extends HasId {
             return new ArrayList<>();
         }
         return outputParameters;
-    }
-
-    @NonNull
-    public List<Tag> getTags() {
-        if (Objects.isNull(tags)) {
-            return new ArrayList<>();
-        }
-        return tags;
     }
 }
