+++ org.planqk.atlas.core/org.planqk.atlas.core.model/src/main/java/org/planqk/atlas/core/model/Tag.java
@@ -1,7 +1,12 @@
 package org.planqk.atlas.core.model;
 
-import javax.persistence.Entity;
+import java.util.List;
 
+import javax.persistence.Entity;
+import javax.persistence.ManyToMany;
+import javax.persistence.MapsId;
+
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,9 +15,21 @@ public class Tag extends HasId {
 
     @Getter
     @Setter
-    String description;
+    String key;
 
     @Getter
     @Setter
-    String name;
+    String value;
+
+    @ManyToMany
+    @Setter
+    @MapsId("algorithm_id")
+    @EqualsAndHashCode.Include
+    private List<Algorithm> algorithms;
+
+    @ManyToMany
+    @Setter
+    @MapsId("implementation_id")
+    @EqualsAndHashCode.Include
+    private List<Implementation> implementations;
 }
