+++ org.planqk.atlas.core/org.planqk.atlas.core.repository/src/main/java/org/planqk/atlas/core/repository/TagRepository.java
@@ -24,10 +24,9 @@ import java.util.List;
 import org.planqk.atlas.core.model.Tag;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
 public interface TagRepository extends JpaRepository<Tag, Long> {
-    List<Tag> findByName(@Param("name") String name);
+    List<Tag> findByKey(String key);
 }
